requirements.txt
Pillow
pandas
googletrans==3.1.0a0
matplotlib
yfinance 
